Fix JWT Role claim issue and refactor profile update endpoint
