+++ EmployeeOnboard.Api/Controllers/AccountsController.cs
@@ -25,9 +25,11 @@ namespace EmployeeOnboard.Api.Controllers
         private readonly IForgotPasswordService _forgotPasswordService;
         private readonly IAuthService _authService;
         private readonly ILogoutService _logoutService;
+        private readonly IUpdateProfileService _updateProfileService;
 
-        
-        public AccountsController(IRegisterService registerService, ILogger<AccountsController> logger, IMapper mapper, IChangePassword changePasswordService, IForgotPasswordService forgotPasswordService, IAuthService authService, ILogoutService logoutService)
+
+
+        public AccountsController(IRegisterService registerService, ILogger<AccountsController> logger, IMapper mapper, IChangePassword changePasswordService, IForgotPasswordService forgotPasswordService, IAuthService authService, ILogoutService logoutService, IUpdateProfileService updateProfileService)
         {
             _registerService = registerService ?? throw new ArgumentNullException(nameof(registerService));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -36,6 +38,8 @@ namespace EmployeeOnboard.Api.Controllers
             _forgotPasswordService = forgotPasswordService;
             _authService = authService;
             _logoutService = logoutService;
+            _updateProfileService = updateProfileService;
+
         }
 
         [HttpPost("register")]
@@ -110,12 +114,8 @@ namespace EmployeeOnboard.Api.Controllers
                     InnerError = innerMessage
                 });
             }
+
         }
-
-
-
-
-       
 
         [HttpPost]
         [Route("logout")]
@@ -131,6 +131,29 @@ namespace EmployeeOnboard.Api.Controllers
             return Ok(result.Message);
         }
 
+        [Authorize(Roles = "Employee, Admin, SuperAdmin")]
+        [HttpPut("update-profile")]
+        public async Task<IActionResult> UpdateProfile(UpdateProfileDTO dto)
+        {
+            var email = User.FindFirstValue(ClaimTypes.Email);
+            var role = User.FindFirstValue(ClaimTypes.Role);
 
+            if (string.IsNullOrEmpty(email) || string.IsNullOrEmpty(role))
+            {
+                _logger.LogWarning("user email or role not found in claims");
+
+                return Unauthorized("Invalid token: missing email or role");
+            }
+
+            var result = await _updateProfileService.UpdateProfileAsync(dto, email, role);
+
+            if (result == null)
+            {
+                _logger.LogError("Failed to update profile for user: {Message}", result.Message);
+                return BadRequest(new { success = false, message = result.Message });
+            }
+
+            return Ok(new { success = true, message = result.Message, data = result.Data });
+        }
     }
 }
